Removed every trace of transformcomponent in favor of having a transform directly on a game entity
First of all
why was a transformcomponent chosen in the first place?
After thinking it through it was kind of redundant, we ALWAYS need a transform, so might as well just make it a vital part of an entity to begin with.

Also had to use entitytransform over Transform, as raylib already had a transform type